# Improve numeric hash function to a better but slightly slower hash function