--- a/.github/regression/micro.csv
+++ b/.github/regression/micro.csv
@@ -9,4 +9,4 @@ benchmark/micro/cast/cast_timestamp_string.benchmark
 benchmark/micro/limit/parallel_limit.benchmark
 benchmark/micro/filter/parallel_complex_filter.benchmark
 benchmark/micro/catalog/add_column_empty.benchmark
-benchmark/micro/groupby-parallel/many_groups_large_values.benchmark
+benchmark/micro/groupby-parallel/many_groups_large_values_small.benchmark