--- a/src/common/checksum.cpp
+++ b/src/common/checksum.cpp
@@ -3,13 +3,17 @@
 
 namespace duckdb {
 
+hash_t Checksum(uint64_t x) {
+	return x * UINT64_C(0xbf58476d1ce4e5b9);
+}
+
 uint64_t Checksum(uint8_t *buffer, size_t size) {
 	uint64_t result = 5381;
 	uint64_t *ptr = (uint64_t *)buffer;
 	size_t i;
-	// for efficiency, we first hash uint64_t values
+	// for efficiency, we first checksum uint64_t values
 	for (i = 0; i < size / 8; i++) {
-		result ^= Hash(ptr[i]);
+		result ^= Checksum(ptr[i]);
 	}
 	if (size - i * 8 > 0) {
 		// the remaining 0-7 bytes we hash using a string hash