Fixes #4208

We move to the following hash function:

```cpp
inline hash_t murmurhash64(uint64_t x) {
	x ^= x >> 32;
	x *= 0xd6e8feb86659fd93U;
	x ^= x >> 32;
	x *= 0xd6e8feb86659fd93U;
	x ^= x >> 32;
	return x;
}
```

See [here](https://nullprogram.com/blog/2018/07/31/) for a description of this hash function.