@Mytherin 2022-08-06
The old hash function was only a multiplication:

```cpp
inline hash_t murmurhash64(uint64_t x) {
 	return x * UINT64_C(0xbf58476d1ce4e5b9);
 }
```

This is very fast to compute, but it does not provide sufficient shuffling of all of the bits in all cases. That is generally fine for normal hash table work, but when doing radix partitioning on specific bits (as we now do in the aggregate hash table, and soon as well in the window & join) all of the bits need to be shuffled sufficiently. Otherwise if the bits you are using for the radix partitioning stay constant the partitioning does not work and you end up with all of the rows going into the same partition (as is illustrated by the linked issue).