@Mytherin 2022-08-07
The new hash function introduces minor regressions in probe-heavy workloads, e.g. Q09 is dominated by hash join probes because of the incorrect join order leading to exploding joins currently, and we see a 15%~ performance regression:

> benchmark/tpch/sf1/q09.benchmark
> Old timing: 1.270279
> New timing: 1.46807

It causes immense speed-ups in cases where the worst case of the old hash function is prevented:

> benchmark/micro/groupby-parallel/many_groups_large_values_small.benchmark
> Old timing: 25.88743
> New timing: 1.625416

Running TPC-H Q09 on my laptop on Tom's branch (see #4274) and the regression in Q09 disappears, as the query is no longer dominated by hash probing:

| Threads | Old Hash | New Hash |
|---------|----------|----------|
| 1T      | 0.37s    | 0.37s    |
| 8T      | 0.06s    | 0.06s    |

We might also consider in the future introducing two different hash functions, and using a hash function that is appropriate for the given purpose (as radix partitioning requires stronger hash functions than simply constructing a hash table). But for now I will merge this PR and accept the minor regressions in favor of stronger robustness of the hash function.